requirements.txt
aiofiles==23.2.1
blinker==1.7.0
click==8.1.7
dlib==19.24.2
exceptiongroup==1.2.1
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pluggy==1.5.0
priority==2.0.0
pytest==8.2.0
Quart==0.19.4
taskgroup==0.0.0a4
tomli==2.0.1
Werkzeug==3.0.1
wsproto==1.2.0
